Устранено дублирование IHostApplication между behaviors
