+++ src/oscript/ExecuteScriptBehavior.cs
@@ -4,25 +4,23 @@ Mozilla Public License, v.2.0. If a copy of the MPL
 was not distributed with this file, You can obtain one 
 at http://mozilla.org/MPL/2.0/.
 ----------------------------------------------------------*/
-using System;
-using OneScript.StandardLibrary;
-using ScriptEngine;
-using ScriptEngine.HostedScript;
+
+using System.IO;
 using ScriptEngine.Hosting;
 using ScriptEngine.Machine;
 using ScriptEngine.Machine.Debugger;
 
 namespace oscript
 {
-    class ExecuteScriptBehavior : AppBehavior, IHostApplication, ISystemLogWriter
+    class ExecuteScriptBehavior : AppBehavior
     {
-        protected string[] _scriptArgs;
         protected string _path;
+        private readonly ConsoleApplicationHost _host;
 
         public ExecuteScriptBehavior(string path, string[] args)
         {
-            _scriptArgs = args;
             _path = path;
+            _host = new ConsoleApplicationHost(args);
         }
         
         public IDebugger DebugController { get; set; } = new DisabledDebugger();
@@ -33,13 +31,11 @@ namespace oscript
 
         public override int Execute()
         {
-            if (!System.IO.File.Exists(_path))
+            if (!File.Exists(_path))
             {
-                Echo($"Script file is not found '{_path}'");
+                _host.Echo($"Script file is not found '{_path}'");
                 return 2;
             }
-
-            SystemLogger.SetWriter(this);
 
             var builder = ConsoleHostBuilder.Create(_path);
             builder.WithDebugger(DebugController);
@@ -50,22 +46,9 @@ namespace oscript
                 builder.Services.RegisterSingleton<ICodeStatCollector>(codeStatProcessor);
             }
 
-            var hostedScript = ConsoleHostBuilder.Build(builder);
-            
+            using var hostedScript = ConsoleHostBuilder.Build(builder);
             var source = hostedScript.Loader.FromFile(_path);
-            Process process;
-            try
-            {
-                process = hostedScript.CreateProcess(this, source);
-            }
-            catch (Exception e)
-            {
-                ShowExceptionInfo(e);
-                return 1;
-            }
-            
-            var result = process.Start();
-            hostedScript.Dispose();
+            var result = _host.RunProcess(hostedScript, source);
 
             if (codeStatProcessor != null)
             {
@@ -76,35 +59,6 @@ namespace oscript
             }
 
             return result;
-        }
-
-        #region IHostApplication Members
-
-        public void Echo(string text, MessageStatusEnum status = MessageStatusEnum.Ordinary)
-        {
-            ConsoleHostImpl.Echo(text, status);
-        }
-
-        public void ShowExceptionInfo(Exception exc)
-        {
-            ConsoleHostImpl.ShowExceptionInfo(exc);
-        }
-
-        public bool InputString(out string result, string prompt, int maxLen, bool multiline)
-        {
-            return ConsoleHostImpl.InputString(out result, prompt, maxLen, multiline);
-        }
-
-        public string[] GetCommandLineArguments()
-        {
-            return _scriptArgs;
-        }
-
-        #endregion
-
-        public void Write(string text)
-        {
-            Console.Error.WriteLine(text);
         }
     }
 }
